+++ knf/validators/network/validators_test.go
@@ -80,6 +80,7 @@ func (s *ValidatorSuite) TestIPValidator(c *C) {
 	})
 
 	c.Assert(errs, HasLen, 1)
+	c.Assert(errs[0].Error(), Equals, "300.0.400.5 is not a valid IP address")
 }
 
 func (s *ValidatorSuite) TestPortValidator(c *C) {
@@ -101,6 +102,8 @@ func (s *ValidatorSuite) TestPortValidator(c *C) {
 	})
 
 	c.Assert(errs, HasLen, 2)
+	c.Assert(errs[0].Error(), Equals, "ABCD is not a valid port number")
+	c.Assert(errs[1].Error(), Equals, "78361 is not a valid port number")
 }
 
 func (s *ValidatorSuite) TestMACValidator(c *C) {
@@ -121,6 +124,7 @@ func (s *ValidatorSuite) TestMACValidator(c *C) {
 	})
 
 	c.Assert(errs, HasLen, 1)
+	c.Assert(errs[0].Error(), Equals, "ABCD is not a valid MAC address: address ABCD: invalid MAC address")
 }
 
 func (s *ValidatorSuite) TestCIDRValidator(c *C) {
@@ -141,6 +145,7 @@ func (s *ValidatorSuite) TestCIDRValidator(c *C) {
 	})
 
 	c.Assert(errs, HasLen, 1)
+	c.Assert(errs[0].Error(), Equals, "127.0.0.1/200 is not a valid CIDR address: invalid CIDR address: 127.0.0.1/200")
 }
 
 func (s *ValidatorSuite) TestURLValidator(c *C) {
@@ -161,6 +166,7 @@ func (s *ValidatorSuite) TestURLValidator(c *C) {
 	})
 
 	c.Assert(errs, HasLen, 1)
+	c.Assert(errs[0].Error(), Equals, "google.com/abcd.php is not a valid URL address: parse \"google.com/abcd.php\": invalid URI for request")
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //