+++ CHANGELOG.md
@@ -1,5 +1,36 @@
 ## Changelog
 
+### 12.44.0
+
+* `[path]` Added method `Compact` for converting path to compact representation
+* `[fmtc]` Added method `Render` for converting color tags to ANSI escape codes
+* `[knf/validators/fs]` Fixed bug with formatting `FileMode` in error messages
+* `[cron]` Improved parsing error messages
+* `[fsutil]` Improved parsing error messages
+* `[cron]` Improved tests
+* `[directio]` Improved tests
+* `[fsutil]` Improved tests
+* `[initsystem]` Improved tests
+* `[jsonutil]` Improved tests
+* `[knf]` Improved tests
+* `[knf/validators/fs]` Improved tests
+* `[knf/validators/network]` Improved tests
+* `[knf/validators/regexp]` Improved tests
+* `[knf/validators/system]` Improved tests
+* `[knf/validators]` Improved tests
+* `[log]` Improved tests
+* `[options]` Improved tests
+* `[req]` Improved tests
+* `[system]` Improved tests
+* `[system/process]` Improved tests
+* `[system/sensors]` Improved tests
+* `[tmp]` Improved tests
+* `[version]` Improved tests
+* `[spellcheck]` Added usage examples
+* `[log]` Fixed bug with rendering messages with colors when `fmtc.DisableColors` set to `true`
+* `[terminal]` Fixed stubs for Windows
+* `[secstr]` Fixed stubs for Windows
+
 ### 12.43.0
 
 * `[terminal]` Added method `ReadPasswordSecure` for reading password into a secure string